switched from int to float for timeouts

+++ tavily/async_tavily.py
@@ -66,7 +66,7 @@ class AsyncTavilyClient:
             include_answer: Union[bool, Literal["basic", "advanced"]] = None,
             include_raw_content: Union[bool, Literal["markdown", "text"]] = None,
             include_images: bool = None,
-            timeout: int = 60,
+            timeout: float = 60,
             country: str = None,
             auto_parameters: bool = None,
             include_favicon: bool = None,
@@ -141,7 +141,7 @@ class AsyncTavilyClient:
                      include_answer: Union[bool, Literal["basic", "advanced"]] = None,
                      include_raw_content: Union[bool, Literal["markdown", "text"]] = None,
                      include_images: bool = None,
-                     timeout: int = 60,
+                     timeout: float = 60,
                      country: str = None,
                      auto_parameters: bool = None,
                      include_favicon: bool = None,
@@ -183,7 +183,7 @@ class AsyncTavilyClient:
             include_images: bool = None,
             extract_depth: Literal["basic", "advanced"] = None,
             format: Literal["markdown", "text"] = None,
-            timeout: int = 30,
+            timeout: float = 30,
             include_favicon: bool = None,
             **kwargs
     ) -> dict:
@@ -237,7 +237,7 @@ class AsyncTavilyClient:
                       include_images: bool = None,
                       extract_depth: Literal["basic", "advanced"] = None,
                       format: Literal["markdown", "text"] = None,
-                      timeout: int = 30,
+                      timeout: float = 30,
                       include_favicon: bool = None,
                       **kwargs,  # Accept custom arguments
                       ) -> dict:
@@ -481,7 +481,7 @@ class AsyncTavilyClient:
                                  include_domains: Sequence[str] = None,
                                  exclude_domains: Sequence[str] = None,
                                  max_tokens: int = 4000,
-                                 timeout: int = 60,
+                                 timeout: float = 60,
                                  country: str = None,
                                  include_favicon: bool = None,
                                  **kwargs,  # Accept custom arguments
@@ -522,7 +522,7 @@ class AsyncTavilyClient:
                          max_results: int = 5,
                          include_domains: Sequence[str] = None,
                          exclude_domains: Sequence[str] = None,
-                         timeout: int = 60,
+                         timeout: float = 60,
                          country: str = None,
                          include_favicon: bool = None,
                          **kwargs,  # Accept custom arguments
@@ -552,7 +552,7 @@ class AsyncTavilyClient:
                                query: str,
                                search_depth: Literal["basic", "advanced"] = "advanced",
                                max_results: int = 5,
-                               timeout: int = 60,
+                               timeout: float = 60,
                                country: str = None,
                                ) -> Sequence[dict]:
         """ Company information search method. Search depth is advanced by default to get the best answer. """
